+++ src/axom/quest/tests/quest_sampling_shaper.cpp
@@ -145,47 +145,6 @@ struct PlaneProjector23
   }
 };
 
-// Utility function to slice a tetrahedron along a plane
-primal::Polygon<double, 3> slice(const primal::Tetrahedron<double, 3>& tet,
-                                 const primal::Plane<double, 3>& plane)
-{
-  primal::Polygon<double, 3> intersectionPolygon;
-
-  // find intersection vertices
-  for(int i = 0; i < 4; ++i)
-  {
-    for(int j = i + 1; j < 4; ++j)
-    {
-      primal::Segment<double, 3> edge(tet[i], tet[j]);
-      double t {};
-      if(primal::intersect(plane, edge, t))
-      {
-        intersectionPolygon.addVertex(edge.at(t));
-      }
-    }
-  }
-  SLIC_ASSERT(intersectionPolygon.numVertices() <= 4);
-
-  // fix the polygon if it bowties
-  if(intersectionPolygon.numVertices() == 4)
-  {
-    // note: using BezierCurve since Axom doesn't currently have intersect(segment, segment)
-    primal::BezierCurve<double, 2> seg1(1);
-    seg1[0] = Point2D(intersectionPolygon[0][0], intersectionPolygon[0][1]);
-    seg1[1] = Point2D(intersectionPolygon[1][0], intersectionPolygon[1][1]);
-    primal::BezierCurve<double, 2> seg2(1);
-    seg2[0] = Point2D(intersectionPolygon[2][0], intersectionPolygon[2][1]);
-    seg2[1] = Point2D(intersectionPolygon[3][0], intersectionPolygon[3][1]);
-    axom::Array<double> sp, tp;
-
-    if(!primal::intersect(seg1, seg2, sp, tp))
-    {
-      axom::utilities::swap(intersectionPolygon[2], intersectionPolygon[3]);
-    }
-  }
-  return intersectionPolygon;
-}
-
 }  // namespace
 
 /// Test fixture for SamplingShaper tests on MFEM meshes
@@ -2032,7 +1991,7 @@ shapes:
     this->initializeShaping(shape_file.getPath());
 
     primal::Plane<double, 3> plane({0, 0, 1}, z);
-    const auto polygon = slice(tet, plane);
+    const auto polygon = primal::slice(tet, plane);
     const double intersectionArea = polygon.area();
     SLIC_INFO(axom::fmt::format("Area of intersection polygon: {}", intersectionArea));